refactor: Login Activity MVP 적용중 login package로 이동 GoalInquiryService의 weeklySchedule 메서드 제거
