+++ app/src/main/java/kr/co/company/capstone/interceptor/TokenAuthenticator.java
@@ -1,13 +1,11 @@
 package kr.co.company.capstone.interceptor;
 
 import android.content.Intent;
-import android.util.Log;
 
 import kr.co.company.capstone.BuildConfig;
 import kr.co.company.capstone.GlobalApplication;
 import kr.co.company.capstone.util.SharedPreferenceUtil;
-import kr.co.company.capstone.activity.LoginActivity;
-import kr.co.company.capstone.dto.ErrorMessage;
+import kr.co.company.capstone.login.LoginActivity;
 import kr.co.company.capstone.dto.login.LoginResponse;
 import kr.co.company.capstone.dto.login.TokenReissueRequest;
 import kr.co.company.capstone.service.LoginService;
