+++ app/src/main/java/kr/co/company/capstone/login/LoginContract.kt
@@ -1,0 +1,4 @@
+package kr.co.company.capstone.login
+
+class LoginContract {
+}
