+++ app/src/main/java/kr/co/company/capstone/login/LoginPresenter.kt
@@ -1,0 +1,4 @@
+package kr.co.company.capstone.login
+
+class LoginPresenter {
+}
