+++ app/src/main/res/layout/social_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.LoginActivity">
+    tools:context=".login.LoginActivity">
 
     <TextView
         android:id="@+id/hello_checker"
